Migrate to Astral tooling; tighten C/Python lint; full TT/TrustZone support
Python tooling migrated to Astral (uv/ruff/ty): flake.nix, pyproject, CI
scripts, pre-commit, and docs now use uv (package mgmt), ruff (lint+format),
and ty (type check) in place of pip/black/mypy.

Lint tightening (all gating, fully green):
- ruff: select = ALL with a curated, documented ignore list; fixed ~1159
  findings incl. real refactors (gdb_server dispatch table; extracted helpers
  in cli/logging/elf_loader; ClassVar; logging.exception; pathlib).
- C compiler warnings: added prototype/declaration hygiene, pointer-arith,
  VLA, double-promotion, -Wformat=2, -Wswitch-default and -Wswitch-enum
  (all enum switches made exhaustive); fixed via shared internal headers
  (exec_handlers.h, decoder_internal.h) and public emu_* decls.
- cppcheck: const-correctness enforced (removed const* suppressions; fixed
  210 test locals + production pointer-const); remaining suppressions are
  documented false-positives/defensive code.
- clang-tidy: WarningsAsErrors gates bugprone/cert/modernize/performance/
  portability.

Full TT/TTA/TTAT support: SAU+IDAU combined attribution, MPU IREGION/IRVALID
and R/RW permission bits, and domain-aware permission gating (consumes the
previously-dead query_state). Adds exported MPU region/permission helpers,
an Executor MPU handle + IDAU state, and C unit tests for TT/MPU.

All green: -Werror build, cppcheck, clang-tidy, C tests (7/7), firmware
tests (43/43), and Python ruff/ruff-format/ty/pytest.

+++ src/arch/armv8m/decoder/decode_thumb32_vfp.c
@@ -358,7 +358,7 @@ int decode_vfp(uint16_t hw1, uint16_t hw2, DecodedInsn *out) {
     /* Not a VFP instruction - generic coprocessor */
     out->type = (BIT(hw1, 4)) ? INSN_MRC : INSN_MCR;
     out->rd = (uint8_t)EXTRACT(hw2, 12, 4);
-    out->imm = ((uint32_t)opc1 << 4) | opc2;
+    out->imm = (opc1 << 4) | opc2;
     return 0;
   }
